Breaks main file down to 4 files: bot, db, models, twitter; employs scoped sessions in sqlalchemy; uses only sqlalchemy vs psycopg2 and sqlalchemy
