--- twitter_reply_bot/templates/index.html
@@ -1,32 +1,0 @@
-{% extends 'bootstrap/base.html' %}
-{% import "bootstrap/wtf.html" as wtf %}
-
-{% block styles %}
-{{ super() }}
-	<style>
-		body { background: #e8f1f9; }
-	</style>
-{% endblock %}
-
-{% block title %}
-Best Movie Actors
-{% endblock %}
-
-{% block content %}
-
-<div class="container">
-	<div class="row">
-		<div class="col-md-10 col-lg-8 mx-lg-auto mx-md-auto">
-			<h1 class="pt-5 pb-2">Welcome to the best movie actors Flask example!</h1>
-
-			<p class="lead">This is the index page for an example Flask app using Bootstrap and WTForms.  Note that only 100 actors are in the data source. Partial names are not valid.</p>
-
-			{{ wtf.quick_form(form) }}
-
-			<p class="pt-5"><strong>{{ message }}</strong></p>
-
-		</div>
-	</div>
-</div>
-
-{% endblock %}
